requirements.txt
azure-functions>=1.18.0
mcp>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
